YOLOv10-BFDA/requirements.txt
torch
torchvision
PyYAML
opencv-python
numpy
